The Ad Size issue is fixed.
Now jump into the next issue I've identified.

For 160x600 and 300x250 we didn't detect the clicktags, and for 300x600 and 728x90 we did detect clicktag, but we need to show a list of the detected clicktag, when detected clicktag isn't a https we need to show that as a warning.
I got this error:
Error: Missing or invalid clickTag implementation.

All the units have this script with the clicktag's URLs:
	<script type="text/javascript">
		var clickTag = "https://www.symbravohcp.com",
			clickTag2 = "http://www.axsome.com/symbravo-prescribing-information.pdf";
	</script>

And later that have listeners that trigger the clicktags

  	<script type="text/javascript">
	    elem('#cta_btn').addEventListener('click', (e)=>{ window.open(window.clickTag)})
	    elem('#fpiLink').addEventListener('click', (e)=>{ window.open(window.clickTag2) })
	    elem('#fpiLink2').addEventListener('click', (e)=>{ window.open(window.clickTag2) })
	    function elem(selectors){return document.querySelector(selectors)};
  	</script>

Also 300x600 and 728x90 give File Structure: Invalid but just give this message:
Error: Invalid file structure. Primary HTML file not found at root.
I'm sure that there are HTML with both of those banners.