I could download the app and sync with Github, also could tested local successfully, followed the Deployment Steps on Netlify but in this last step at the end of the Deployment Netlify give me an URL: https://html5-validator.netlify.app/
but it returns:
Page not found
Looks like you’ve followed a broken link or entered a URL that doesn’t exist on this site.

If this is your site, and you weren’t expecting a 404 for this path, please visit Netlify’s “page not found” support guide for troubleshooting tips.

Give me an advise about step to review to fix this.