I need to install a plugin, th instructions as follow:
Install via HTML
Include this line at the end of your website's body, just before the </body> closing tag. If you can only place code in the head, that's fine too.

<!-- 100% privacy-first analytics -->
<script data-collect-dnt="true" async src="https://scripts.simpleanalyticscdn.com/latest.js"></script>
<noscript><img src="https://queue.simpleanalyticscdn.com/noscript.gif?collect-dnt=true" alt="" referrerpolicy="no-referrer-when-downgrade"/></noscript>

The App is already uploaded on Netlify, Which and where is the file that a I need to copy this?